peripherals/shdwc: cosmetic changes

+++ drivers/peripherals/shdwc.c
@@ -67,39 +67,36 @@ union _shdwc_cfg {
  *        Exported functions
  *----------------------------------------------------------------------------*/
 
-void shdwc_configure_wakeup_mode (uint32_t config)
+void shdwc_configure_wakeup_mode(uint32_t config)
 {
 	union _shdwc_cfg cfg;
 
 	cfg.uint32_value = SHDWC->SHDW_MR;
 
-	cfg.bfield.lpdbcen0 = (config & SHDW_MR_LPDBCEN0_ENABLE)? 1:0;
-	cfg.bfield.lpdbcen1 = (config & SHDW_MR_LPDBCEN1_ENABLE)? 1:0;
+	cfg.bfield.lpdbcen0 = (config & SHDW_MR_LPDBCEN0_ENABLE) ? 1 : 0;
+	cfg.bfield.lpdbcen1 = (config & SHDW_MR_LPDBCEN1_ENABLE) ? 1 : 0;
 	cfg.bfield.lpdbc = (config & SHDW_MR_LPDBC_Msk) >> SHDW_MR_LPDBC_Pos;
-	cfg.bfield.rttwken = (config & SHDW_MR_RTTWKEN)? 1:0;
-	cfg.bfield.rtcwken = (config & SHDW_MR_RTCWKEN)? 1:0;
-	cfg.bfield.accwken = (config & SHDW_MR_ACCWKEN)? 1:0;
-	cfg.bfield.rxlpwken = (config & SHDW_MR_RXLPWKEN)? 1:0;
+	cfg.bfield.rttwken = (config & SHDW_MR_RTTWKEN) ? 1 : 0;
+	cfg.bfield.rtcwken = (config & SHDW_MR_RTCWKEN) ? 1 : 0;
+	cfg.bfield.accwken = (config & SHDW_MR_ACCWKEN) ? 1 : 0;
+	cfg.bfield.rxlpwken = (config & SHDW_MR_RXLPWKEN) ? 1 : 0;
 	cfg.bfield.wkupdbc = (config & SHDW_MR_WKUPDBC_Msk) >> SHDW_MR_WKUPDBC_Pos;
 
 	SHDWC->SHDW_MR = cfg.uint32_value;
 }
 
-void shdwc_set_wakeup_input (uint32_t input_enable, uint32_t input_type)
+void shdwc_set_wakeup_input(uint32_t input_enable, uint32_t input_type)
 {
-	uint32_t wuir = (input_enable & 0x0000FFFF) | (input_type&0xFFFF0000) ;
+	uint32_t wuir = (input_enable & 0x0000FFFF) | (input_type & 0xFFFF0000);
+
 	SHDWC->SHDW_WUIR |= wuir;
 }
 
 void shdwc_do_shutdown(void)
 {
-	SHDWC->SHDW_CR = (uint32_t)((SHDW_CR_KEY_PASSWD) | SHDW_CR_SHDW);
+	SHDWC->SHDW_CR = SHDW_CR_KEY_PASSWD | SHDW_CR_SHDW;
 }
 
-/**
- * \Get Status
- *
- */
 uint32_t shdwc_get_status(void)
 {
 	return SHDWC->SHDW_SR;
